c++: convert sys/tracefd and sys/shell_menu modules to C++
Convert modules to C++:
- sys/tracefd
- sys/shell_menu

lib/obstacles module is partially ported because it was needed to compile
applications.

All files using these modules have been converted from .h/.c to .hpp/.cpp
and adapted to use the new C++ interface of converted modules.
This does not mean that they are converted in C++, it remains mainly C code
included in C++ files.

+++ controllers/common_controller/include/ctrl.hpp
@@ -146,7 +146,7 @@ typedef struct {
 
     const uint16_t blocking_speed_treshold;         /**< Blocking speed
                                                          treshold */
-    const uint16_t blocking_speed_error_treshold;   /**< Blocking speed error
+    const double blocking_speed_error_treshold;   /**< Blocking speed error
                                                          treshold */
     const uint16_t blocking_cycles_max;             /**< Blocking cycles
                                                          maximum number */
